[audio_core] Avoid pipeline underflow on startup
Problem: when starting up a device, we perform some audio effect
inintialization. For effects that may take 'a long time' to initialize,
this causes the audio device to get behind on some of the initial frames
it needs to produce for the ring buffer. This in turn may result in the
first mix job needed to mix more than the expected 10ms, and combining
with the deadline profile this can lead to an pipeline underflow.

Solution: include the current driver safe write-ptr when computing what
inital frame should be sent. This results in us trimming those initial
frames (not mixing) instead of attemting to do the work to mix that
backlog.

Bug: 62830
Bug: 64368
Multiply: audio-core-api-pipeline-tests
Multiply: audio-core-google-pipeline-tests-astro
Change-Id: I0d65c2b067f37a48b5a494dda619598089a96813
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/425663
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>
Testability-Review: Tom Bergan <tombergan@google.com>