[settings] Play max volume earcon on voice query
The earcons agent was only set up to play the max volume earcon
if the volume was increased via hw. This refactors so that sw changes to
max volume will also play the earcon if it is changed to max. This
eliminates the need for the media button events as the changes will all
come through the switchboard.

Fixed: 42952
Test: fx test -o setui_service_tests
Multiply: setui_service_tests

Change-Id: Iaae7d114315f62dc9404645b758bd761a8b7bc91
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/453344
Commit-Queue: Chip Fukuhara <cfukuhara@google.com>
Reviewed-by: Bryce Lee <brycelee@google.com>
Testability-Review: Bryce Lee <brycelee@google.com>