[components] Log open of non-existent path
If a component tries to open a path in a service directory in its
namespace which does not exist, log that attempt. Possibly the
component author forgot to include a `use` declaration. This is
an important debugging affordance whose true positives we think
pay for the false ones.

Fixed: 64179

Change-Id: Id7a2de7676d5a149fc681b882f9d9a4c866ae117
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/453407
Commit-Queue: Justin Mattson <jmatt@google.com>
Reviewed-by: Adam Lesinski <adamlesinski@google.com>
Testability-Review: Adam Lesinski <adamlesinski@google.com>