[audio_core] After a pipeline underflow, log the time spent on CPU
Often, PIPELINE UNDERFLOWs are due to the mixer thread getting
interrupted by the CPU, such as when the audio mixer code has been paged
out, or when some other Zircon scheduling decision causes extra latency.

To identify these causes of UNDERFLOW separate from our mixing code
simply running for too long, this CL logs the total amount of time the
mixer thread spent actually running on the CPU. If this is low relative
to the wall-clock time of the mix job, we can infer that the UNDERFLOW
is due to a scheduling problem, not due to the mixer running for too
long.

Bug: b/172857417
Bug: 62830
Change-Id: Ie227fff371b693a0ea3a33b61872170e610d4939
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/453348
Reviewed-by: Payton Bradsky <pbradsky@google.com>
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Testability-Review: Payton Bradsky <pbradsky@google.com>
Commit-Queue: Tom Bergan <tombergan@google.com>