[msd-vsi-vip] Kill the context if SubmitCommandBuffer fails.
Currently we have no way of knowing whether a batch was successfully
completed or not, as we just drop it if SubmitCommandBuffer fails (which
triggers the expected semaphore). This should help with root
causing the current flake, where it seems like ResetAfterSubmit
is actually failing but not being caught by the test.

BUG=60383

Change-Id: Ib145631568cadc3deec2c96a6456dfdbc1fa97c5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/453585
Reviewed-by: Craig Stout <cstout@google.com>
Testability-Review: Craig Stout <cstout@google.com>
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>