[ddk][logging] Take file and line as parameters to zxlogvf
The file and line number currently associated with zxlogvf is always
incorrect. This is because it always points to the function implementing
the varargs. It needs to be propogated from the caller in order to be
correct

Change-Id: I231b3f19da65e69cc50e31dfe1cdf50c1f994579
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/455161
Testability-Review: Suraj Malhotra <surajmalhotra@google.com>
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Yilong Li <liyl@google.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Craig Stout <cstout@google.com>
Commit-Queue: Suraj Malhotra <surajmalhotra@google.com>