[testing] Add method to look for bag of strings
Add a method to emulator which looks for a group of strings in
any appearance order. This is useful in situations where the test
writer wants to assert that multiple things happened, but the
order of their occurrence is not deterministic, perhaps because
they happen in different processes or threads.

Change-Id: I7a21c905e7298e1fccd5b52b768a739dee95ace7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/455168
Reviewed-by: Petr Hosek <phosek@google.com>
Reviewed-by: Scott Graham <scottmg@google.com>
Commit-Queue: Justin Mattson <jmatt@google.com>
Testability-Review: Scott Graham <scottmg@google.com>