[docs][drivers] Mention the two sides of filtering
Explicitly call out the fact that there is both a producer and consumer
side to logging. This is a design decision of syslog, and some users are
unaware of it.

How to control the consumer-side of filtering should be elaborated upon
in the diagnostics logs, as it is general functionality of syslog.

Change-Id: I847ad9cce59a4d172796663a77eb3a752bdb26d3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/455659
Reviewed-by: Alex Legg <alexlegg@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Abdulla Kamar <abdulla@google.com>