[bt][a2dp] Add ConnectedPeers::connected_stream
Add a stream to the ConnectedPeers struct which produces a peer whenever
one is connected. This allows for responses to peers being connected.

Use the new stream to add peers to the ControllerPool for AVDTP testing
when they are connected, and remove the plumbing used to previously
add the peers.

Bug: 61530
Test: added bt-a2dp::connected_peers::test::connect_notifies_streams
Test: fx test bt-a2dp-unittests bt-a2dp-sink-tests bt-a2dp-source-tests
Change-Id: Ibeb036c44f82adcc932733232128cd5079c1bf89
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/454977
Commit-Queue: Marie Janssen 💖 <jamuraa@google.com>
Reviewed-by: Jeff Belgum <belgum@google.com>
Testability-Review: Jeff Belgum <belgum@google.com>