[blobfs] Fix some hangs possible when encountering write errors.
When blobfs encounters write errors it was possible for it to hang. It's
not really something we need to support on real devices because there's
no graceful recovery if this happens, but it is an issue where we want
to do power-failure tests. I have tests that follow that exercise this.

Change-Id: I3970a63855fe2dd70c5edf35a703bcb8815b5860
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/455902
Reviewed-by: James Sullivan <jfsulliv@google.com>
Testability-Review: James Sullivan <jfsulliv@google.com>
Commit-Queue: Chris Suter <csuter@google.com>