[SetUI] Address flaky task manager test.
This fix addresses an issue comparing the beginning of
the longest running active task to the beginning of the
test. Due to the precision of the timestamps, it is
possible that the spawned task's start will not be after
the test start timestamp. However, it should not occur
before.

Fixed: 65040
Test: fx test setui_service_tests
Change-Id: I5c90b14a993d17b752bceab10c20434e7722feca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/456474
Reviewed-by: Chip Fukuhara <cfukuhara@google.com>
Reviewed-by: William Xiao <wxyz@google.com>
Testability-Review: Chip Fukuhara <cfukuhara@google.com>
Testability-Review: William Xiao <wxyz@google.com>
Commit-Queue: Bryce Lee <brycelee@google.com>