[fx][emu][gn sdk] Fix femu.sh failing to use KVM, fix internal `fx emu` dependency
femu.sh relies on `-a auto` which uses internal fx state - not present
in the femu context - in order to resolve this option for hardware
acceleration. Fall back to providing that bit of state ($HOST_CPU) if it
is not in the environment at the time of execution. Previously this was
causing KVM to never be used, and the emulator would run a lot slower.

Bug: 62631
Fixed: 64194
Change-Id: Ide30594d7e0b8fe86e7a4c3707a6eeb33571c52e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/453783
Reviewed-by: Renato Mangini Dias <mangini@google.com>
Reviewed-by: Wayne Piekarski <waynepie@google.com>
Commit-Queue: Wayne Piekarski <waynepie@google.com>
Testability-Review: Wayne Piekarski <waynepie@google.com>