[component_manager] Introduce CapabilityTypeName.
Introduce a new type to cm_rust CapabilityTypeName that provides a
bounded enum for the previous &'static string that was returned from
CapabilitySource.type_name(). By converting this type to an enum it can
be more efficiently compared by the PolicyChecker in the hot-path of
capability routing without incurring the cost of a full string compare.
It also prevents having to deal with the unbounded possible values of a
string when the type_name is infact bounded to several types.

Bug: 61866
Change-Id: If1c13468641c156abde762afefbf38c1b4a09c3f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/453824
Reviewed-by: Gary Bressler <geb@google.com>
Testability-Review: Gary Bressler <geb@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Benjamin Wright <benwright@google.com>