[wlan] Add IesMerger for scan results
To prepare for consolidating the IEs in BssDescription, we need a way to
merge IEs blocks from a beacon and probe response into a single IEs block.
This commit introduces the IesMerger struct for such purpose.

IesMerger keeps an underlying buffer of IEs bytes and an index of IEs
it's going to keep. When it's merging new IEs block, it loops through
the block and keeps the new IE if it doesn't already have such IE type or
if the new IE contains more information. When IesMerger decides to keep
an IE, it appends the IE into the buffer and updates the index. Later,
when the merge operation is finalized by the caller, IesMerger would use
the index to construct the merged IEs block from the underlying buffer.

Bug: 63004
Test: fx test wlan-common-tests
Change-Id: Idf4d95d8514ccb5693f3b6779d6c3466589750a3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/453964
Commit-Queue: Kiet Tran <kiettran@google.com>
Reviewed-by: Eric Wang <eyw@google.com>
Testability-Review: Eric Wang <eyw@google.com>