[wlan] Fix the seed of the RC4 keywrap algorithm
The seed of the RC4 algorithm used by 802.11 appends the encryption
key to an IV. This is not explicitly specified for RC4 use in general,
but this is in fact the seed construction used for RC4 in WEP. See
IEEE 802.11-2016, Section 12.3.2.4.3 - WEP seed construction.

In practice, our RC4 keywrap algorithm is only used for
WPA2-TKIP. This is because we punt WEP encryption to the driver and
WPA1 does not typically encrypt EAPOL keyframes.

Fixed: 65175, 63888
Test: Both `fx test wlan-rsn-tests` and ACTS test
WlanSecurityComplianceABGTest pass.

Change-Id: Ie2766a6be2684066e3143db40567f3864546720b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/454199
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>
Reviewed-by: Dylan Swiggett <swiggett@google.com>
Testability-Review: Dylan Swiggett <swiggett@google.com>