[blobfs] Integrate the external decompressor.
Integrates the sandboxed decompressor and adds options for enabling it
which default to off for now. Adds metrics to count how often the
decompressor is used to the read stats.

This makes a change to greatly simplify the SeekableRemoteDecompressor
as the previous interface could not integrate properly given the way
that the pager places data in the vmos.

Bug: 56464,62177
MULTIPLY: external-decompressor-runner (fuchsia): 20

Change-Id: I284c484d0e6354dfca2f5844e10187246c4a0398
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/455355
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Testability-Review: James Sullivan <jfsulliv@google.com>