[artifactory] Upload GSYM to GCS
GSYM is a more efficient format for debug symbols that we'd like to
start using. In order to do that we need to upload GSYM debug symbols in
artifactory, to both the legacy "debug" and the new "buildid"
namespaces.

This will be a no-op until we start setting an `output_gsym` arg on some
builds and populating the "gsym" field in binaries.json to point to each
binary's GSYM file.

Fixed: 61677
Change-Id: Ie712fbd8c75a8838395afa9045db0b9e9ac11cd9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/456547
Testability-Review: Oliver Newman <olivernewman@google.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Anthony Fandrianto <atyfto@google.com>