[kernel][mexec] Avoid allocation overlap
Passes the end of the target kernel image location to the identity
allocator to avoid overlapping allocations. This is neccessary since
the destination for the kernel image is not selected by the pmm and so
it otherwise does not know how to avoid it.

Without this the included asserts can trip, and without the asserts
this overlap can result in mexec failure, typically manifesting as a
triple fault on x86.

Change-Id: I1fff01431795555db7213a55381d5ba8ef52b0e9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/455901
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: David Greenaway <dgreenaway@google.com>
Testability-Review: David Greenaway <dgreenaway@google.com>