[musl] Don't use empty initializers when building for C
In C, it is technically illegal to initialize a struct by saying
something like

  MyStruct my_struct = {};

This said, GCC will allow this by default, as will Clang. If, however,
someone is building using Clang, and -Wgnu-empty-initializer is
passed, this code will generate a warning.

Currently, the definitions of some of the pthreads structure
initializers in our version of MUSL make use of this empty-initializer
extension, meaning that 3rd party code might run afoul of this
technically illegal pattern, depending on their build settings.

So, avoid this issue by initializing with { 0 } instead of {}.  Note
that this will not work in C++, nor is using designated initializers
an option as those are not officially supported by C++ until C++20
(and we currently only require C++14).  To work around this, two
versions of the initializer are provided, using the __cplusplus
preprocessor macro to choose which to use.

Test: new unit test
Fixed: 64794
Change-Id: I379eeaecec6d3980657cce6749b60444cd69dde8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/456734
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Testability-Review: Roland McGrath <mcgrathr@google.com>
Commit-Queue: John Grossman <johngro@google.com>