[dma-buffer-test] Use |options| parameter in zx_vmo_create
In the refactor to make dma-buffer-test use lib/fake-object a
mistake was made that led to the options field being ignored
in the proxied zx_vmo_create. Since the only option available
to that syscall is for resizing it wasn't a problem in the tests,
but this fixes it.

Bug: 53888
Change-Id: Ia39127c696e41dbdb4404ba64fd9590ec7ffc15d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/457236
Testability-Review: Christopher Anderson <cja@google.com>
Commit-Queue: Christopher Anderson <cja@google.com>
Fuchsia-Auto-Submit: Christopher Anderson <cja@google.com>
Reviewed-by: David Dorwin <ddorwin@google.com>