[audio_core][tests] Pipeline fidelity tests, pt.2
This is the second part of a refactor to the hermetic
audio testing library, to add fidelity testing. This
needed to be landed in stages, because this lib is used in
other repositories. With this CL, the refactor is complete.

This CL also contains additional renaming not done in the
initial CL.

Test: test-only code, no product change
Multiply: audio-core-google-pipeline-tests-astro
Multiply: audio-core-google-pipeline-tests-sherlock

Change-Id: I36b005ec0ce22d8d87a5d59872d991a80b4698c5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/453183
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>
Testability-Review: Tom Bergan <tombergan@google.com>