[brcmfmac] Fix glitch in scanning during AP start
When starting an AP the driver would abort any on-going scans. If a scan
was initiated while the AP was being started that scan would also be
rejected. This was done to speed up AP start. However there was a small
window of time where a scan would be allowed in between an on-going scan
being aborted and the flag that prevented future scans being set.
Frequently SME would issue a new scan as soon as the previous one was
aborted and this would occur before that flag was set, allowing the scan
to go through.

This change moves the setting of the flag to before the scan is aborted
so that it is guaranteed to be set when SME receives the event
indicating that the scan was aborted. This prevents the retry scan from
starting.

BUG: b/173283408
BUG: 65069
Test: Rapid scan attempts while AP is starting
      sim_fw tests pending in other CL
Change-Id: I155271e8e0b581f83b45e462a1fb1978f84e6682