[ffx] rebind mdns sockets on interface cycle
If an interface on the host cycles, the interface would miss due to a
hash key mismatch with the port number on the interface.

Fixing this issue highlighted new issues in testing:
Apple devices with touchbars have a misconfigured persistent IPv6
interface associated with the touchbar iBridge. This interface is
now explicitly ignored when running on macOS hosts, by relying on
the fact that it always has the same link local IPv6 address.

Coming out of standby, some interfaces may cycle, such as USB
interfaces attached to external hubs. In order to combat side effects
of the associated interface reconfigurations, particularly with eager
DHCP trickery, it is necessary to reduce the default discovery
frequency, as the first send may not actually make it onto the wire
under these conditions, in a manner that is racy and thus hard to
test even with manual interaction.

Lock contention is slightly reduced by reducing the use and scope of
the socket tasks lock in discovery. This can help in cases where
socket binds take an unusual amount of time, another condition that
is possible to trigger with repetitive manual testing.

Although it is rarely observed on target platforms that the recv_from
call fails (in practice only the sends observe loss), that case is
now handled gracefully, and terminates the query receive loop,
enabling the socket to be re-established on the next discovery pass.

In order to reduce the network cost associated with the increased
discovery frequency, discovery is switched to unicast discovery. This
is in-line with the behavior of service browsing as observed from
mdnsresponder, as invoked by device-finder.

Change-Id: Ic50645175cdcaf747eaa1e3e6275fdb1ae441a37
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/456538
Commit-Queue: James Tucker <raggi@google.com>
Testability-Review: James Tucker <raggi@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>