[e2e][intl] Increases wait between retries from 1 to 3s
Increases wait time between probes for timezone change in the timezone
test from 1s to 3s.

This change is based on an assumption that flakes due to long message
latency are simply long latency tails, and not a more complicated issue.
The change is a harmless way to disprove the long latency assumption: if
the flake remains, we need to dig deeper, but if it is resolved, then
we got a quick and easy fix.

Bug: 65318
Change-Id: Ia95302d7a22b3fdbaa2057e22440a6745ce39f00
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/457798
Testability-Review: Filip Filmar <fmil@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Reviewed-by: Viktar Donich <viktard@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>