[gigaboot] Raise initial EFI allocation from 6 MB -> 8 MB
gigaboot requests a fixed allocation for the kernel load
zone; the allocation size was a constant, which was large
enough for plain kernels but no longer large enough for
kASAN kernels when global instrumentation was added
(6.5 MB vs 6 MB).

Raise the initial allocation request to 8 MB to cover
this for now.

Fixes boot of kASAN + global instrumentation kernel on NUC.

Tested:
* Booted --variant kasan kernel + zedboot on Intel NUC6i5
* Booted plain variant on GCE; a --variant kasan kernel
  will not work on GCE, either before this CL or after.
  GCE can only support a <= 2.5 MB EFI allocation at our
  load address.

  GCE continues to work because of a fallback in the code;
  if the initial allocation fails, we step down to a
  smaller size.

A real fix would be to make the kernel relocatable or at
least parse the kernel to find the appropriate zone size
and request that. This is just a targeted fix to make the
kasan trybots work again.

Bug: 30033 Kernelmode ASAN for Zircon
Bug: 58770 KASAN Fails to boot on NUC
Bug: 32374 Make x64 kernel relocatable
Bug: 32223: Gigaboot's pre-allocation of the kernel range gets
Change-Id: I028d9a95c9fdab4eeae46653f25efa7320ffe054
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/421177
Commit-Queue: Venkatesh Srinivas <venkateshs@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Testability-Review: Roland McGrath <mcgrathr@google.com>