[SetUI][Audio] Use flag to signal change.
The Audio controller currently uses a timestamp to signify
a stream's volume has changed. This logic depends on
the precision of the timestamps being greater than the
time between requests. This turns out not to be true in
some environments, such as tests. This change converts
the signal to a flag, where an alternation between high
and low indicates a change in the stream's volume.

Fixed: 65747
Test: fx test -o setui_service_tests
Multiply: setui_service_tests
Change-Id: Ia18998e072c40b3992a5e86f19883b4456d30bc0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/457672
Commit-Queue: Bryce Lee <brycelee@google.com>
Reviewed-by: Tom Robinson <robinsontom@google.com>
Testability-Review: William Xiao <wxyz@google.com>