[zbitl] Add a fuzzer for copy codepaths
Several bugs were found around bounds-checking and with the
copy-creation byte range method (specifically, with the
creation-CopyWithSlop method failing to return the proper slop value in
the non-clone case and failing to zero out `to_offset`-many bytes in
the created storage). To address the latter, we updated the behaviour
of the Create() trait to do the zeroing out of the head of the storage.

Fuzz support for the storage copy codepaths will be added in a
follow-up.

Bug: 62638
Change-Id: I45df27421dc34b52afdabb5d20197123167a8f06
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/456117
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: David Greenaway <dgreenaway@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Testability-Review: Roland McGrath <mcgrathr@google.com>
Testability-Review: David Greenaway <dgreenaway@google.com>