[wlan][sme] Handle bss_desc in scan and connect.
Populates the bss_desc field in scan results, and updates SME
to bypass the join scan and connect directly when passed a
bss_desc in a connect request.

This is the final CL to support connection by BSSID in core. We
will revisit this logic if/when we make connection by BSSID the
only supported emthod.

Bug: 65481
Test: Updated/new unit tests

Change-Id: I3c0c7940caf41568a9a90244ef54a33c5fcae27b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/452513
Commit-Queue: Dylan Swiggett <swiggett@google.com>
Reviewed-by: Charles Celerier <chcl@google.com>
Testability-Review: Charles Celerier <chcl@google.com>