[kernel][sched] Fix an edge case in changing to a deadline proifle.
An edge case was introduced when adding support for heterogenous
processor performance. In that change, deadline tasks are given
proportionally more time when running on slower processors to
ensure they get the same effective amount of work done before the
deadline. Because scale factors were added to various calculations,
a debug assert was added the common reschedule path to ensure that
reaching the target preemption time is consistent with the scaled
time slice expiring.

When changing from a fair profile to a deadline profile, the
remaining time slice is clamped to the capacity of the profile to
prevent violating bandwidth isolation, while still permitting the
thread to run when there is time remaining. If the thread is
running at the time of the profile change (e.g. it is setting its
own profile), then the target preemption time computed from the
unscaled time slice when it was a fair thread becomes stale and
will assert when compared to the scaled time slice, due to being
a deadline thread now.

Fixup the target preemption time to account for the performance
scale of the CPU it is running on when switching to a deadline
profile.

Bug: 61508
Test: Looping fair <-> deadline on a little core does not assert.
Change-Id: I671ba0eec06d6eb4ea84ed41a3e0537a63f17a5d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/458020
Reviewed-by: John Grossman <johngro@google.com>
Testability-Review: Corey Tabaka <eieio@google.com>
Commit-Queue: Corey Tabaka <eieio@google.com>