[fdio] Implement ftruncate for memfd
This implementation is somewhat racy because we need to touch a few
different aspects of the VMO to resize it accurately. We have a similar
problem in memfs. It's likely we'll need a new kernel API for doing this
work atomically, but we can refactor this code to use that API when its
available.

Bug: 60236
Change-Id: I7d9d914dab5f1a261730fb1ab87b9d86ce84091c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/457948
Commit-Queue: Adam Barth <abarth@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Testability-Review: James Robinson <jamesr@google.com>