[blobfs] Support page requests larger than 256MB
Add support to page in requests for ranges larger than the transfer
buffer and the decompression buffer. zx_vmo_op_range with
ZX_VMO_OP_COMMIT on the entire blob generates a page request spanning
the whole blob. The transfer and decompression buffers are sized at
256MB, which limits the size of blobs that can be committed entirely to
256MB. This change removes that limitation. The user pager now reads in
and decompresses the range in chunks, so that they can be accommodated
in the buffers.

In practice this code path is not expected to be exercised often, since
it only applies to blobs larger than 256MB.

Test: blobfs-unittest

Bug: 64454
Change-Id: I7334767aef9caccca65467a552009d39f593ae06
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/456735
Reviewed-by: James Sullivan <jfsulliv@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Testability-Review: James Sullivan <jfsulliv@google.com>
Testability-Review: Chris Suter <csuter@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>