[mac][vulkan][build] Fix libvulkan.dylib symlinking
Previously the dylibs were implemented as:
1. hardlink prebuilt/.../libvulkan.1.dylib -> out/.../libvulkan.1.dylib
2. softlink out/.../libvulkan.dylib to libvulkan.1.dylib

This didn't work with gn stamp files because step 2 touches a stamp
file, but when the timestamp of its input is looked up, the link is
followed back to libvulkan.1.dylib, and that in turn shares a hardlink
mtime with the dylib in prebuilt (which isn't modified so depends on
when cipd synced it.)

As a result, the mtime of the input libvulkan.dylib is always "older"
than the .stamp file, so the build is always dirty.

To resolve this, change to a plain copy from prebuilt -> out, and in the
same step, generate the symlink. Remove the symlink.py script and make
it a custom helper for this purpose as it's (as evidenced here) a bit of
a footgun in GN/ninja.

Bug: 65872
Change-Id: Ib3cdc82e9317860de5e2ed16ddd44e836c6fa2cc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/458287
Reviewed-by: Yilong Li <liyl@google.com>
Reviewed-by: David Turner <digit@google.com>
Fuchsia-Auto-Submit: Scott Graham <scottmg@google.com>
Testability-Review: Scott Graham <scottmg@google.com>
Commit-Queue: Scott Graham <scottmg@google.com>