[bt][bt-gap] Errors restoring host bonds do not fail Commit()
When processing the fuchsia.bluetooth.sys.bootstrap.Commit() function,
do not return an error if storing the data in the store succeeds but
restoring the data to an active host fails. Instead, this is treated as
a Host error but not an error with the commit procedure.

In order to test behavior correctly, also implement an in-memory version
of the bt-gap Store that doesn't use the Stash component or protocol,
and so can easily be unit-tested. Migrate existing tests to this
implementation and remove the `stub()` function that previously existed.

Bug: 63013

Test: fx test bt-gap-unittests

Change-Id: I934fbf968df80e26591a6b4de7f52ef2e6b5b691
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/455597
Commit-Queue: Nick Pollard <nickpollard@google.com>
Reviewed-by: Lucas Jenkins <lucasjenkins@google.com>
Testability-Review: Lucas Jenkins <lucasjenkins@google.com>