[hlcpp] Remove Message class
This is part of FTP-028 and an LSC to migrate from the Message
class to HLCPPIncomingMessage and HLCPPOutgoingMessage.

This is the last step of this LSC -- to remove the Message
class. "fx lsc presubmit" has been run on this CL and it should
not cause breakages.

Bug: 41920
Fixed: 65523

Change-Id: Ifd4cbe7bd4f17c097634f68cf68dd1425a304e19
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/457616
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Testability-Review: Yifei Teng <yifeit@google.com>
API-Review: Adam Barth <abarth@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>