[fidlc] Fix crash on nullable bit fields.
Bits should behave identically to enums when used as nullable fields
in structs; that is, the compiler should error and inform the user that
this behavior is banned.

This reporting was not happening for bits, and was resulting in a crash
instead, due to a missing case in the switch statement that validates
that a token is nullable.  This case, along with a test for this bug,
are added in this changelist.

Fixed: 40994
Test: fx test fidl-compiler; fx test fidlc_golden_tests
Change-Id: Iee76841f4cf77a05841ca2f54c70e6a6d1ad3b67
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/458345
Commit-Queue: Alex Zaslavsky <azaslavsky@google.com>
Reviewed-by: Felix Zhu <fcz@google.com>
Testability-Review: Felix Zhu <fcz@google.com>