[firmware] flesh out README
Adds more detail to the README, such as details of each included
component and general porting instructions.

We will be adding this to the Firmware SDK shortly, so this will be the
main reference for developers to port it to new bootloaders.

Change-Id: I7954b455af5bb28212618b1ca949f0afa18fcc8a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/458336
Fuchsia-Auto-Submit: David Pursell <dpursell@google.com>
Reviewed-by: Yecheng Zhao <zyecheng@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>