[netstack] Do not store endpoint with hard error
Do not add an endpoint with a hard error state to the endpoints map.
The endppoint is removed from the map only on an EventHUp, and such
endpoints would no longer see an EventHUp.
For ex: Accept of an endpoint that was reset post established state.

Fixed: 66033
Test: fx test -v netstack_test -- -test.run=TestTCPEndpointMapAcceptAfterReset

Change-Id: Ia059ba5b0101b012bbec498563b8b6ea0f986047
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/459335
Commit-Queue: Mithun Iyer <iyerm@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Testability-Review: Tamir Duberstein <tamird@google.com>