[bt][a2dp] CodecNegotiation prefers a direction
Now that we are presented with the opportunity for connecting both
source and sink endpoints, we need to choose which direction to connect
when we have both.

Add a preferred direction to the CodecNegotiation, and prefer the
correct direction in all of the A2DP profiles. We currently prefer
to be Source in bt-a2dp.

Bug: 61530
Test: fx test bt-a2dp-lib-tests
Change-Id: Ie87ce7dc0287b8905c426b428d710709c226664e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/455495
Commit-Queue: Nicolas Sylvain <nsylvain@google.com>
Testability-Review: Jeff Belgum <belgum@google.com>
Fuchsia-Auto-Submit: Marie Janssen 💖 <jamuraa@google.com>
Reviewed-by: Jeff Belgum <belgum@google.com>