[debugger] Add missing C++ operators.
Fills out the operator enum sufficiently to allow all "operator X"
definitions to be parsed. Most of the operators were in-place update
operators like "++" and "+=". These are added but unimplemented and will
give an error message. They are uncommon to use in the debugger.

The main missing operator was ~ which required some special handling to
disambiguate from destructor names. Becuase the debugger doesn't always
have the correct context and a better implementation seemed
unnecessarily complex for this uncommon operator, ~ followed by a word
is always treated as a name. To express a binary not, the caller needs
to add a space or another token like a paren.

Bug: 49698

Change-Id: I4d7d384dee1b5d86c40ff7bdeecf59c7bba6ecf1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/458454
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Dangyi Liu <dangyi@google.com>
Testability-Review: Dangyi Liu <dangyi@google.com>