[session] Implement add command in session_control
session_control add <element spec> should add an element to the
session, if the session exposes the ElementManager protocol.

Ideally the session manager would expose the session's element manager
directly in the session manager cml.

If this is not possible, another way of doing this is to have the
session_manager component expose the ElementManager, just like it
exposes fuchsia.session.Launcher, and proxy incoming requests to the
session.

Can be closed when session_control can be used to launch an element in a session.

Bug: 47125
Change-Id: I48e177054253c626aebac4b5884f7d18ac1ae436
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/436207
Commit-Queue: Casey Borders <casbor@google.com>
Reviewed-by: Mukesh Agrawal <quiche@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Testability-Review: Kevin Lindkvist <lindkvist@google.com>
Testability-Review: Yegor Pomortsev <ypomortsev@google.com>