[kernel][x86] Multi-page support in MappingCursor
This extends the MappingCursor implementation to support multiple
physically discontiguous pages instead of just a single contiguous one.
In supporting this the MappingCursor class is extended to be more
object like and move all its data to be private with accessors.

The additional paddr support is not taken advantage of in this commit,
and this change should be perfectly semantic preserving.

Test: Covered by existing kernel unit tests.

Change-Id: Ic39ad410742ae401e2c42307faeb74231691bc3a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/458265
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Venkatesh Srinivas <venkateshs@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Testability-Review: Rasha Eqbal <rashaeqbal@google.com>